tests/eas/rfc.py: Fix path of ondemand/sampling_rate file
In case a target has only one frequency domain, the ondemand/sampling_rate
file is located under /sys/devices/system/cpu/cpufreq/ondemand and not
under the individual cpu directory.

Fix this by checking if the ondemand/sampling_rate file exists in the
individual cpu directory after the ondemand governor has been set.

If this is not the case write sampling rate to the system global one.

Tested on JUNO and Hikey target.

Signed-off-by: Dietmar Eggemann <dietmar.eggemann@arm.com>