libs/utils/env: move EM data into board specific files
The main target.config file is overloaded of information and complex to read.
This patch aims at simplify it a bit by moving some details.

Specifically, this patch introduces a new folder containing platform specific
data, currently located under libs/utils/platforms.
This folder can contain a JSON file for each of the supported boards.
For the time being we use the board file to host the default (simplified)
energy model definition.

The patch remove the definition of the EM from the main target.config file.
However, we keep the possibility to override the default EM by specifing
one on the main target.config file.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>