libs/utils/conf: Make the json parser cope with trailing commas
The configuration files are written using JSON but the standard JSON
doesn't allow trailing commas in dicts or arrays, which is very annoying
when enabling and disabling entries.  Some JSON parsers do silently
ignore trailing commas, but that is not the case for the python
implementation.  Add a simple regexp to remove commas at the end of
arrays and dictionaries.

This is a fix for issue #4