libs/utils/env: speedup reboot time
Once a target is rebooted, currently we wait for a time interval which is
configured to be long enough to allows all boards the actually reboot.
Sometimes that time can be too short or even unnecessary long.

This patch adds the support for a PING based detection of the same target
being back online. The maximum timeout is increased to be 2 minutes because
in general we expect a proper detection of the target before this timeout.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>