tests/eas/rfc: fixup governor setup after reboot
After a board reboot the target specified governor is not configured if we
have:
1. multiple configurations all using the same governor
2. the governor required by these configurations is not the default one
   used by the board at boot time

This patch fixes that issue by removing the "optimization" code which was
used to track at host side the governor we (assume) be enabled in the target.

Every time the cpufreq setup function is called we enforce the configuration
of the governor required by the current target configuration.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>