base: make parsed_data a generator
With big traces, the parsed_data array becomes huge (up to gigabytes).
parsed_data is only used as a parameter for pandas.DataFrame.  pandas
accepts a generator instead of a list.  Therefore instead of creating a
huge array for pandas, pass a generator to pandas and let it process the
incoming information as it needs to.