libs/utils/env: add suppor to load test configuration from file
At TestEnv initialisation, the target configuration can be specified either
inline or as a JSON file to load. The test configuration is instead forced
to be inlined.

This patch align test configuration handling to target configuration.
The required code to load the test configuration from a file is added
and the test configuration is now tracked with the self.test_conf
local variable.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>