libs/utils/env: move modules definition into board specific section
The specification of devlib modules to load into the target and/or
test configuration file is a not portable solution, especially when
we need to run the same test on different boards which do not support the
same set of modules, e.g. SMP platform does not support "bl" module.

This patch update the modules configuration by moving their definition
into the specific board configuration section of the TestEnv intialization
code.
Each board can now list the *complete* list of modules which are supported
and only these will be initialized.

Still the target and test specific configuration file allows to add
additional modules eventually not loaded by a specific board configuration.

The list of modules being loaded is now clearly reported as a log message.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>