libs/utils/env: add support to exlude a list of modules
Boards defined a *complete* list of supported modules. However, sometimes
for a specific experiment we do not need all of them.

This patch adds the required support to specify, either in the target
or test specific configuration files, a list of devlib modules which we do
not want to be loaded. This allows for example to avoid loading "hwmon" on
JUNO/TC2 in case we are not interested in energy measurements.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>