ipynb: add an example notebook to plot kernel function profiling data
Kernel functions profiling support allows to collect statistics on time
spend whiting a set of specified kernel function.
Profiling data are collected into a JSON file "trace_stat.json" in the
same output folder where the trace of an experiment is generated.

This patch provides a simple notebook to process and plots all the profiling
data collected during some experiments. This should be just good to have
an initial visual representation of the overheads. However, a proper
evaluation can require a specific setup of the experiments. For example
it could be required to focus on a single CPU by pinning the workload there
as well as to use the proper CPUFreq governor. All these details are left
to the user expertise.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>