libs/utils/env: add support to specify a test results folder
Currently, each time we create TestEnv object we create a new results folder
which has a timestamp based name. That approach is not flexible enough when
for example we want to run multiple notebooks which produces experiments
for the same "test case". Moreover, sometimes could be useful to have a more
meaningful name for a set of tests.

This patch adds the possibility to specify an "ID" for a test configuration.
This ID, when specified, is used to name a sub-folder of "results" which
will contains all the output of the test.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>