README.md: add note about ipython version to be installed
Some IPython Notebooks from the Quickstart tutorials are written in JSON
nbformat version 4. The problem is that if you install ipython and
ipython-notebook using apt-get you will get version 1.2.1-2 (as of today) which
does not support JSON nbformat version 4. Hence, the solution is to remove
ipython and ipython-notebook and install them using pip instead.