libs/utils/executor: add a generic tests execution engine
A tests executor is a module which support the execution of a
configured set of experiments. Each experiment is composed by:
- a target configuration
- a worload to execute

The executor module can be configured to run a set of workloads
(wloads) in each different target configuration of a specified set
(confs). These wloads and confs can be specified by the "tests_config"
input dictionary.

All the results generated by each experiment will be collected a result
folder which is named according to this template:
     results/<test_id>/<wltype>:<conf>:<wload>/<run_id>
where:
- <test_id> : the "tid" defined by the tests_config, or a timestamp
              based folder in case "tid" is not specified
- <wltype>  : the class of workload executed, e.g. rtapp or sched_perf
- <conf>    : the identifier of one of the specified configurations
- <wload>   : the identified of one of the specified workload
- <run_id>  : the progressive execution number from 1 up to the
              specified iterations

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>