ipynb: executor: add example usage of the executor module
The libs/utils/executor.py module provides a simple yet useful execution
engine to run a reconfigured set of experiments.
This patch adds an example Notebook which shows how to configure and
use the execution engine.

Comments in the notebook should provide enough information to understand
how to use that engine from within a Notebook to run a set of experiments.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>