libs/wlgen: update log statements formatting to match execution engine one
During tests execution, the output of the execution engine is going to be
intermixed with the output of the workload generator. Right now the two
modules uses a different formatting template which makes a bit difficult
to read the generated log.

This patch updates the formatting string of the workload generator modules
to match that of the execution engine.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>