libs/utils/env: add ping_time parameter in reboot
When rebooting, the reboot function repeatedly pings the target.
However, rebooting might require some seconds and in order to avoid
pinging the machine before it actually reboots, we can set the
ping_time option in target.config to tell the function how long to
wait before pinging.

Also, a comment has been added in target.config to specify what the
file represents and to highlight that commented parameters are
optional.