libs/utils/trace: use standard mainline events name
Most recent versions of TRAPpy exposes all the available events as attributes
of the Run object which have the same name of the corresponding tracepoint
event.

This patch is a first of a series which aims at align the LISA interface
to the new TRAPpy naming conventions. The first patch simply rename all
the events of interests in LISA to use the same name of the tracepoint.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>