libs/utils/trace: replace trace_data usage with new TRAPpy interface
Most recent versions of TRAPpy exposes all the available events as attributes
of the Run object which have the same name of the corresponding tracepoint
event.

This patch remove the not more required trace_data mapping of tracepoints
to convenient names and switch to the usage of the more generic Trace.df()
method which allows to get the DataFrame corresponding to one of the parsed
tracepoints by just specifying its name.
An exception is raised in case the event is not available in the trace.
Thus, client code should better check with Trace.hasEvents() before trying
to get a DataFrame.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>