libs/utils/trace: add support for overutilized events
The sched_overutilized tracepoint is used to report when the system is
running in EAS mode or it's over the tipping point.

This patch adds the required code to parse overutilized events as well as
to compute and report the percentage of time the system was running in
overutilized mode (for the specific trace).

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>