libs/utils/filters: return tasks dictionary for all filters
Some filter functions allows to select a list of tasks which matches some
search criteria. These functions returns now a list of PIDs, which is not
usable for trace_analysis defined functions.

Update the filter functions to return a dictionary of "name:pid" for all
the tasks selected by that functions. This result can be used to plot other
graphs defined by trace_analysis functions.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>