submodules: tracking ARM-Software versions of TRAPpy and BART
This patch update the external libraries to track the official mainline
repos for BART and TRAPpy. We target a specified branch/tag version.

The init_env script has been updated as well to ensure that, for example
in case we change the remote repository, all the submodules are properly
initialized before the update.

This version of init_env should also work properly in zsh.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>