tests: re-factoring to use the new Executor module
The current implementation of the RFC tests collects tests from EAS,
SchedFreq and SchedTune under a single module. This solution does not
scale and makes the code unnecessary difficult to read.
Moreover, the data collection step is common to all the test and can be
replaced from a standard module.

This patch redefine the layout of the currently available tests to be more
rational as well as more easily extensible. Each framework has its own
folder and the data collection step is now implemented using the newly
introduced Execution module.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>