libs/utils/executor: fix workload pinning on CPUs for non bL systems
When running on a non big.LITTLE system, pinning tasks based on CPUs type
specifiers like "bigs_first" is not possible because of the missing "bl"
module in devlib.
However, still we want to be able to run workloads defined using these
attributes.

This patch maps CPUs specifiers on non bL systems using this convention:
[bigs|littles]_first: fist available CPU
[bigs|littles]_last: last available CPU

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>