libs/utils/env: make reboot more robust by improving host resolution
reboot has been made more robust by reading the ARP cache before
shutting down the target so that we save target IP and MAC address.
Then, reboot tries to ping the target using the IP, but in case the
IP changes it will try to resolve the MAC address and use the new IP
instead.