plotter: StaticPlot: move legend below graph
Legend has been moved below the graph for better readability and because
it sometimes overlapped with figure title.

Labels in the legend can now span 3 columns, i.e. up to 3 labels are
shown horizontally inside the legend. If there is a fourth label, it
will wrap to the a new row.

In addition, 'borderaxespad=0.' is needed to avoid the legend being
cut off from the figure.