plotter: StaticPlot: make_title show trace name only for single trace plots
The make_title function now prints the trace name only when needed, that
is when a single trace is being plot. In fact, in this case there will
be no legend and it is good to let the reader know which trace is shown
by putting its name in the title. On the contrary, where there are more
traces, the legend is generated and therefore putting names in the title
becomes redundant.