libs/utils/executor: ensure each workload is executed in the proper CGroup
When the target is configured to use CGroups, if required, by default all
workloads are executed on the default group. Each workload has than the
chance to override the default group by specifying a different one.

This patch ensure that the default CGroup is used for the execution of
a workload only if it is not overridden by a local configuration.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>