plotter: ILinePlotGen: substitute cols parameter with per_line
The per_line attribute is written into the json file that describes the
plot to be displayed. It identifies the number of figures in a line and
therefore the number of columns. For this reason the cols parameter has
been deleted from the ILinePlotGen class.

This also solves the issue with too narrow plot when plotting a single
figure. Previously, per_line was always set to the default value equal
to 2 which caused the plot of a single figure to be in the first column
even if no figures are being plotted in the next one.

close #134