plotter: always show lower centered legend
For both static and interactive plot show the legend below the plots and
centered with respect to the area occupied by the figures.

Static plots will show three labels per line in the legend by default.
This default value can be changed by setting the legend_ncol parameter
to the desired value. Notice also that the legend is shown in any case even
for a single trace. Since the trace name will now always appear in the legend,
in the next commits it will be removed from the title.

Interactive plots will show the legend with a single label per line, as there
is no option in javascript to tell the maximum number of labels that can be
printed in one line.