ipynb: profiling: update example to generate data from notebook
This patch makes use of the executor module to generate all the data
required for the profiling of a set of functions. In this what the notebook
is self-contained and it's possible to run it directly to profile a
configured set of kernel functions.

The functions to profile are defined, as part of the FTrace configuration,
by this attribute:
  test_conf::ftrace::functions

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>