plotter: fix constraint filtering with multiple filters and pivots
Constraint._apply() generates a boolean array to filter the data frame
based on the filters.  If a ValueError is raised we try to remove the
duplicated indexes before retrying to apply the filter.

When we have multiple filters, by constantly applying the criterion on
"values" we keep reducing it, but by keeping the criterion the same size
as the original dataframe ("data") we force reindexes that make pandas
raise a ValueError.  For example:

                criterion = values.map(lambda x: True)
                for key in self._filters.keys():
                    if key in data.columns:
                        criterion = criterion & data[key].map(
                            lambda x: x in self._filters[key])
    -->                 values = values[criterion]
                result[AttrConf.PIVOT_VAL] = values
                return result

We try to fix this in handle_duplicate_index() but there is no need to
do that.  We just have to stop applying the full criterion on reduced
dataframes.  The most efficient way to do it is to build the criterion
for the full constraint (filters and pivot) and then apply it only once
on the dataframe.