plotter: Remove handling duplicate indexes from Constraint
After 9fe317fdb547 ("plotter: fix constraint filtering with multiple
filters and pivots") it is not necessary to handle duplicate indexes in
Constraint any more as we don't need to reindex dataframes.  Kill this
code from Constarint.

Note that this does not kill handle_duplicate_index from trappy, there
are still valid uses in ILinePlot and the grammar.