grammar: apply a window to data accesses
Parser() operations happen across the whole trace.  While it is possible
to filter traces before feeding them to Parser (e.g.
trappy.FTrace(window=(1, 2))) the truth is that sometimes you need to
parse the trace to know the window in which you want to apply the test.
Instead of forcing the user to reload the trace after calculating the
window, let Parser() know that it should operate on a small window of
trace instead.