plotter: create uniform title style for every type of plot
A uniform style for printing plots titles is added. The final figure
will show the following titles:

 - Master Title: empty by default. This is the title of the entire
                 plot. Shown at the top center of the figure if the user
                 provides it.

 - Pivot/Column Title: title of each plot is empty by default. It shows
                       pivot names if any pivot is specified by the user.
                       Otherwise, it shows column names if permute is set
                       to True.

For interactive plots the title given by the user was ignored and
therefore not shown in the plot. Therefore, the proper attribute has
been added to the class and it's now printed in the figure.