libs/wlgen/rta: refactor RTA methods to be class objects
Currently we generate specific workloads patterns by using a member function
of the RTA class which returns a list of phases to execute. This approach
has two main disadvantages:
1. the embedded documentation is not exposed to Notebooks
2. phases generated by a class cannot be easily "composed" to define a
   more complex workload

This patch refactor these RTA methods to be standalone classes. This allows
to get embedded documentation available on Notebooks and prepare the ground
to introduce workloads shapes composition. To this purpose, all the shape
classes are now derived from a _TaskBase class. That class provides the
new _TaskBase::get() method which can be used to get the dictionary
used to configure RTApp.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>