Trigger: narrow the data_frame with the pivot before applying the filter
Some complex filters require you to look at previous values of the
data.  For example, we may want to know when a frequency is no longer
used for a given cpu.

Trigger(trace, trace.cpu_frequency, pivot="cpu",
        filters={"frequency": freq_no_longer(max_freq)}, value=-1)

freq_no_longer(max_freq) returns True if the previous frequency of this
cpu was max_freq and the current is not.  The tricky bit is "of this
cpu".  With the current implementation, all freq_no_longer() sees is the
previous value, but it doesn't know to which cpu it corresponds.

Filter the data_frame with the pivot (as we are going to end up doing
regardless) before applying the filter to avoid exposing the filter to
values that are useless to it.