libs/wlgen/rta: add support for modular RTA phases
For some experiments it could be useful to generate workload which
change over time and that can be easily defined as a composition of
"basic shapes". The RTA workload generator allows already to defined
some common shapes (e.g. Periodic, Pulse, Ramp) but its missing the possibility
to compose these shapes to create more complex patterns.

This patch add this functionality by allowing to concatenate tasks definitions
to generate a single Task with variable behaviors along time.

An example of this new composition feature is provided by the sched_freq
smoke test notebook, which has been update to use the new API to express
the phases required by the test. For example, the light-heavy-light task
is described as:

  light1 = Periodic(duty_cycle_pct=10, duration_s=1.0, period_ms= 10,
                    cpus=str(target.bl.bigs_online[0]))
  heavy1 = Periodic(duty_cycle_pct=90, duration_s=0.1, period_ms=100)
  light2 = Periodic(duty_cycle_pct=10, duration_s=1.0, period_ms= 10)

  lhl_task = light1 + heavy1 + light2

  rtapp.conf(kind='profile', params={'task1': lhl_task.get()})

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>