plotter: StaticPlot: fix data plotted when selecting permute option
The correct behaviour when permute is set to True is to show one plot per
trace and in each plot display the data corresponding to all the columns
specified by the user.

The data is represented in a specific axis based on computations on the index
of each (constraint, pivot) pair. In order to show the right traces in the
correct plot, we order the constraints based on two keys: the column from
where to get the data, and the name of the trace.