Trigger: allow filters to be any callable
The current test for the filters only checks whether the value of the
filter is a function.  Actually, we don't care what it is, as long as we
can call it.  As somebody said in StackOverflow, we're looking for
duck-typed objects because we want the ones that quack, not the ones
that fit in a duck-sized container.  I don't care if you're a rubber
duck, if you can quack, I'll apply you to the data frame.