libs/utils/test: fix multiple run of the executor
Tests executions and data collection is driven by the Executor module which
should be executed just one time to run all the experiments required by
the following tests.
The current implementation of LisaTest creates the Executor object
and run its from within the ctor method of the base class. This ctor is
called by nosetests before each test runs. This imply that we currently run
data collection multiple times.

This patch fixes this issue by switching to the usage of the setUpClass
class method, which is granted to be called just one time by nosetest.
This mehod is used in each test to call the LisaTest::_init() method
which is now the one in charge to initialized and execute tests.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>