libs/utils/env: fix host resolution for target connected via WiFi
When the remote target is connected via WiFi, the ARP cache cannot be used
to resolve an IP to a MAC address.
This patch adds a special case for host defined via IP. In case the MAC
address cannot be found in the ARP cache we do not generate an error but
just mark as "UNKNOWN" the MAC address of the remote host.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>