libs/utils/executor: make setting governor's tunables generic
Instead of hardcoding governors' names when changing tunables we can
used devlibs's methods to achieve the same in a generic way.

Do it. Ideally we would also like to add a for_each_policy() method to
devlib instead of iterating over all target's online cpu. But, that
belongs to a different patch.

Signed-off-by: Juri Lelli <juri.lelli@arm.com>