ILinePlot: create uniform indexes for mupltiple traces
When plotting multiple traces with different indexes (i.e. x-axis values) the
legend of the interactive plotter shows one label at a time because for the
other traces no index (and as a consequence no value) exist.

To avoid this, we merge the indexes of the traces to get a common x-axis for
all of them and then fill the resulting NaN values accordingly to the selected
drawstyle, that is:

 - default, interpolation

 - steps-pre, copy previous value (bfill)

 - steps-post, copy next value (ffill)

 - steps-mid, copy next value (ffill)

Merging is performed right before the data is added to the plot. This makes
sure that "only" the data to be plotted is affected by the merge. In fact, at
this point filtering has already been done.