ftrace: match comm names that have a '[' as part of their name
990f925224107 ("run: update special fields regexp to parse task names with a
space") fixed parsing task names that had spaces by assuming that no task
would have a "[" as part of their name.  The day has come that we have found
a task with a "[".

Set the regexp to match anything up to 16 characters.  The kernel seems to
limit the task name to 16 characters so this should be safe from now on (until
we find another task that breaks the regexp, that is).