ipynb: tutorial: add notebooks and other assets
This patch adds a first version of a complete tutorial covering most of
the LISA modules. The initial tutorial: 00_LisaInANuthshell provides an
overall view of the main LISA features and, for each of the main
modules, a link to another notebook which details some more advanced
usages of the API provided by that module.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>