libs/utils/trace_analysis: fix overutilized bands computation
It could happens that multiple events are generated (e.g. on different CPUs)
at the exact same time. When this happens, the current code to compute
overutilization bands generates bands which are described by a list of
intervals thus confusing the following plotting code.

This patch ensure that we do not have 0 length bands, which are not required
by the plotting code. This allows to have a proper representation of
overutilization intervals with the maximum available granularity.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>