tools/scripts/cgroup_run_into: fix execution in root CGroup
When a task is required to run in the root CGroup (i.e. "/") the
current code fails and actually allocate the task in the deepest
CGroup of each controller.

This patch adds a special case to identify cgroup.procs files of
each controller and move the task on these groups. The existing
code is still valid for the execution of a task on a specified
subgroup (i.e. "/subgroup/complete/path") but has been updated
a little bit to be more robust and aligned with the root CGroup
case.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>