libs/wlgen/workload: allows FTrace collection also when running in background
When a workload is started in background we do not currently properly support
the collection of FTrace. An FTrace collection is started but actually not
stopped from the WlGen class.

This patch allows to start an FTrace collection even when the task runs in
background. To achieve that goal we need however to specify a suitable
value for end_pause_s, which will be used to define for how long to trace
the system while the application is running in background.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>