sched: cope with SysTrace's sched_switch event
For trace-cmd traces, we use the raw trace to parse sched_switch.
That's not possible for SysTrace traces, we don't have access to the raw
data.  Luckily SysTrace doesn't mangle the data as much as trace-cmd's
plugin, so a simple string replace is enough to get rid of a stray
"==>".