libs/utils/env: add N5X (bullhead) target support
The bullhead kernel does not properly report big.LITTLE CPUs in
the output of /proc/cpuinfo. However, since this is a well known
device it is safe to provide a custom initialisation of the
platform module for that target.

This patch defineds the big.LITTLE layout of a Nexus 5X device which
can be used by specifying:
  board = "n5x"
in your TestEnv configuration file.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>