plotter: allows to parse systrace generated traces
The trace plotter exposes an API which allows to plot a trace not yet parsed.
Howevever, thanks to the support added by PR #160, which allows to collect
ftrace events using SysTrace and the atrace agent, an input trace for the
trace plotter can also have been generated usign systrace.

This patch ensure that the trace plotter can be used also for SysTrace
parsed trace events. In case a trace filename is specified in input, we
relay on the extension of the file to guess the trace format.
Systrace generates by default HTML files, thus an input trace named with
the "html" extension will be parsed using the SysTrace parser.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>