libs/utils/trace: provide a meaningful error if we don't understand the trace format
Up until 34f5d28c5ef4 ("libs/utils/trace: add support for systrace trace
format") we treated every trace path as a file captured with trace-cmd,
so we always parsed it with with trappy.FTrace and let it fail if it
wasn't.  Now that we have support for SysTrace as well, we barf with a
weird error message if we don't understand the trace_format.

Raise an exception instead to make it easier to understand what is going
wrong.