ftrace: fix default name for traces
Having "." as default name for traces was preventing TRAPpy from
assigning a meaningful name, like "Trace 0", to traces for which the
user does not explicitly set a name. As a consequence, every plot with
such traces was showing a dot in the legend and in the title (where
trace name is relevant).

close #168