libs/utils/env: disable rootfs remount as RW partitions
Most of the targets provides a predefined RW partition where it is safe
to install LISA/devlib required components.
This patch remove the forced remount of the root partition which is not
supported on some devices. For example it's not easy to get a RW root
partition on Android >=N devices.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>