plotter: EventPlot: drop newline when reading css
When EventPlot embeds the css into the notebook, it adds an extra "\n"
for every line.  css_fh.readlines() returns all the lines as an array
but the lines already contain a "\n" at the end of the line.  The
"\n".join() afterwards adds another "\n", so you end up with an extra
newline for every line.  Drop the "\n" when reading the css so that it
is embedded without empty lines in between every line in the notebook.